+++ units/src/lib.rs
@@ -34,7 +34,7 @@ pub mod block;
 pub mod fee_rate;
 pub mod locktime;
 pub mod parse;
-pub mod timestamp;
+pub mod time;
 pub mod weight;
 
 #[doc(inline)]
@@ -43,6 +43,6 @@ pub use self::{
     amount::{Amount, SignedAmount},
     block::{BlockHeight, BlockInterval},
     fee_rate::FeeRate,
-    timestamp::BlockTime,
+    time::BlockTime,
     weight::Weight
 };
